// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EstadoVerificacion = {
  SI: 'SI',
  NO: 'NO',
  Pendiente: 'Pendiente'
} as const

export type EstadoVerificacion = (typeof EstadoVerificacion)[keyof typeof EstadoVerificacion]


export const EstadoContacto = {
  Pendiente_de_Verificacion: 'Pendiente_de_Verificacion',
  Activo: 'Activo',
  Inactivo: 'Inactivo'
} as const

export type EstadoContacto = (typeof EstadoContacto)[keyof typeof EstadoContacto]


export const EstadoCuota = {
  Pendiente: 'Pendiente',
  Vencida: 'Vencida',
  Pagada: 'Pagada',
  En_Acuerdo: 'En_Acuerdo'
} as const

export type EstadoCuota = (typeof EstadoCuota)[keyof typeof EstadoCuota]


export const EstadoSolicitud = {
  Pendiente: 'Pendiente',
  Aprobada: 'Aprobada',
  Rechazada: 'Rechazada',
  Expirada: 'Expirada'
} as const

export type EstadoSolicitud = (typeof EstadoSolicitud)[keyof typeof EstadoSolicitud]


export const PrioridadSolicitud = {
  Baja: 'Baja',
  Media: 'Media',
  Alta: 'Alta',
  Critica: 'Critica'
} as const

export type PrioridadSolicitud = (typeof PrioridadSolicitud)[keyof typeof PrioridadSolicitud]